requirements.txt
flake8==6.1.0
importlib-metadata==6.8.0
iniconfig==2.0.0
mccabe==0.7.0
mypy==1.7.0
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyQt6==6.6.0
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
pytest==7.4.3
shiboken6==6.6.0
tomli==2.0.1
types-Pillow==10.1.0.2
typing_extensions==4.8.0
yapf==0.40.2
zipp==3.17.0
